sjendoceni vetvi

+++ Bakalarka-cityRework/GenetickyAlgoritmus/Cities.cs
@@ -34,21 +34,27 @@ namespace GenetickyAlgoritmus
             int index1 = getIndex(city1);
             int index2 = getIndex(city2);
 
-            int x1 = Convert.ToInt32(Controller.allCitiesTable.Rows[index1]["X"].ToString());
-            int x2 = Convert.ToInt32(Controller.allCitiesTable.Rows[index2]["X"].ToString());
-            int y1 = Convert.ToInt32(Controller.allCitiesTable.Rows[index1]["Y"].ToString());
-            int y2 = Convert.ToInt32(Controller.allCitiesTable.Rows[index2]["Y"].ToString());
+            double x1 = Convert.ToDouble(Controller.allCitiesTable.Rows[index1]["X"].ToString());
+            double x2 = Convert.ToDouble(Controller.allCitiesTable.Rows[index2]["X"].ToString());
+            double y1 = Convert.ToDouble(Controller.allCitiesTable.Rows[index1]["Y"].ToString());
+            double y2 = Convert.ToDouble(Controller.allCitiesTable.Rows[index2]["Y"].ToString());
 
             return Math.Round(Math.Pow(Math.Pow(x2 - x1, 2) + Math.Pow(y2 - y1, 2), 0.5), 4);
 
         }
 
+        /// <summary>
+        /// Vrátí vzdálenost mezi 2 městy. Vstupem jsou index měst
+        /// </summary>
+        /// <param name="city1"></param>
+        /// <param name="city2"></param>
+        /// <returns></returns>
         public static double getDistance(int city1, int city2)
         {
-            int x1 = Convert.ToInt32(Controller.allCitiesTable.Rows[city1]["X"].ToString());
-            int x2 = Convert.ToInt32(Controller.allCitiesTable.Rows[city2]["X"].ToString());
-            int y1 = Convert.ToInt32(Controller.allCitiesTable.Rows[city1]["Y"].ToString());
-            int y2 = Convert.ToInt32(Controller.allCitiesTable.Rows[city2]["Y"].ToString());
+            double x1 = Convert.ToDouble(Controller.allCitiesTable.Rows[city1]["X"].ToString());
+            double x2 = Convert.ToDouble(Controller.allCitiesTable.Rows[city2]["X"].ToString());
+            double y1 = Convert.ToDouble(Controller.allCitiesTable.Rows[city1]["Y"].ToString());
+            double y2 = Convert.ToDouble(Controller.allCitiesTable.Rows[city2]["Y"].ToString());
 
             return Math.Round(Math.Pow(Math.Pow(x2 - x1, 2) + Math.Pow(y2 - y1, 2), 0.5), 4);
         }
@@ -73,6 +79,11 @@ namespace GenetickyAlgoritmus
             return coordinate;
         }     
         
+        /// <summary>
+        /// Vrátí ID města, když dodám název města
+        /// </summary>
+        /// <param name="city"></param>
+        /// <returns></returns>
         public static string getId(string city)
         {
             int index = getIndex(city);
@@ -80,6 +91,12 @@ namespace GenetickyAlgoritmus
 
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="city1"></param>
+        /// <param name="city2"></param>
+        /// <returns></returns>
         public static double getDistanceId(string city1, string city2)
         {
             int index1 = getIndexId(city1);
@@ -94,6 +111,11 @@ namespace GenetickyAlgoritmus
 
         }
 
+        /// <summary>
+        /// Vrátí index města podle názvu
+        /// </summary>
+        /// <param name="city"></param>
+        /// <returns></returns>
         public static int getIndexId(string city)
         {
             int coordinate = 0;
@@ -104,6 +126,11 @@ namespace GenetickyAlgoritmus
             return coordinate;
         }
 
+        /// <summary>
+        /// Vrátí název města podle indexu
+        /// </summary>
+        /// <param name="i"></param>
+        /// <returns></returns>
         public static string getCityName(int i)
         {
             return Controller.allCitiesTable.Rows[i]["Obec"].ToString();
